modified:   demo/ethernet/ethernet.c 	modified:   demo/ethernet/lib/print.c 	modified:   my/disk/lib/else.c 	modified:   my/jarvis/dowhat/f4.s 	modified:   my/jarvis/init/prepare.s

+++ my/disk/lib/else.c
@@ -58,6 +58,6 @@ void str2data(BYTE* str,QWORD* data)
 	for(i=0;i<8;i++)
 	{
 		if(str[i]<=0x20) break;
-		*data+=(str[i])<<(i*8);
+		*data+=((QWORD)str[i])<<(i*8);
 	}
 }
